pyproject.toml
[tool.poetry]
name = "framegrab"
version = "0.3"
description = "Easily grab frames from cameras or streams"
authors = ["Groundlight <info@groundlight.ai>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
opencv-python = "^4.4.0.46"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"